style: lint sources with latest 1.69 toolset

+++ pallets/circuit/src/machine/test_extra_stress.rs
@@ -132,7 +132,7 @@ pub fn setup_xtx_with_10_sfx(
             sfx_arr_of_10[sfx_index as usize]
                 .generate_id::<circuit_runtime_pallets::pallet_circuit::SystemHashing<Runtime>>(
                     &local_ctx.xtx_id[..],
-                    sfx_index as u32,
+                    sfx_index,
                 ),
         );
     }
@@ -182,7 +182,7 @@ pub fn bid_for_n_out_of_10_sfx_in_xtx(
             .input
             .generate_id::<circuit_runtime_pallets::pallet_circuit::SystemHashing<Runtime>>(
             &local_ctx.xtx_id.0,
-            sfx_index as u32,
+            sfx_index,
         );
 
         assert_ok!(Machine::<Runtime>::compile(
